Check string replacements everywhere
Check string replacements everywhere, not just in .resx files. This way `XlfDocument` need not have a dependency on `ResxDocument`.
