+++ src/XliffTasks/Model/XlfDocument.cs
@@ -155,26 +155,23 @@ namespace XliffTasks.Model
                     changed = true;
                 }
 
-                if (sourceDocument is ResxDocument)
+                // If the source and target require different numbers of formatting items then reset
+                // the target string completely. This avoids problems when the source has been updated
+                // to remove formatting items--when formatting the target string we won't have as many
+                // replacement items as it calls for, leading to an exception.
+                // And if the source string is updated to use _more_ items then formatting with the
+                // target string is likely to produce misleading (or outright meaningless) text. In
+                // either case we lose nothing by just reverting the string until it can be localized
+                // again.
+                var sourceReplacementCount = sourceElement.Value.GetReplacementCount();
+                var targetReplacementCount = targetElement.Value.GetReplacementCount();
+
+                if (targetReplacementCount != sourceReplacementCount)
                 {
-                    // If the source and target require different numbers of formatting items then reset
-                    // the target string completely. This avoids problems when the source has been updated
-                    // to remove formatting items--when formatting the target string we won't have as many
-                    // replacement items as it calls for, leading to an exception.
-                    // And if the source string is updated to use _more_ items then formatting with the
-                    // target string is likely to produce misleading (or outright meaningless) text. In
-                    // either case we lose nothing by just reverting the string until it can be localized
-                    // again.
-                    var sourceReplacementCount = ResxDocument.GetReplacementCount(sourceElement.Value);
-                    var targetReplacementCount = ResxDocument.GetReplacementCount(targetElement.Value);
+                    targetElement.Value = sourceNode.Source;
+                    stateAttribute.Value = "new";
 
-                    if (targetReplacementCount != sourceReplacementCount)
-                    {
-                        targetElement.Value = sourceNode.Source;
-                        stateAttribute.Value = "new";
-
-                        changed = true;
-                    }
+                    changed = true;
                 }
 
                 // signal to loop below that this node is not new
